% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Visualizing probabilities under the normal curve for Stats250}
\usage{
prob()
}
\description{
\code{prob} assists students in plotting probabilities under the normal
curve. When entering input, be sure you're entering the number corresponding
to the selection you want (rather than typing in the selection itself). Type
in \code{prob()} to the command line to get started!
}

