% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{Visualizing P-values in hypothesis testing for Stats250}
\usage{
pval()
}
\description{
\code{prob} assists students in plotting/visualizing p-values in hypothesis
testing. When entering input, be sure you're entering the number
corresponding to the selection you want (rather than typing in the selection
itself). Type in \code{pval()} to the command line to get started!
}

